Do not raise on missing objects.
The use of ActiveRecord's #find meant missing IDs raised AR::RecordNotFound. This might come up if an object was created and then deleted without its trigram data also being deleted. To avoid the exception this uses #where instead of #find which will ignore missing results.

Due to the method with which the records are sorted to match the trigram query, in these cases nils were cropping up in the results, so this uses #compact to just return the records.

The first test case ensures no exceptions are raised. The second test case ensures the fuzzy find still works and that no nils are returned.

Conflicts:
	lib/fuzzily/searchable.rb